backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the AI Impostor Game"
authors = [
    {name = "AI Software Engineer"},
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "firebase-admin",
    "python-dotenv",
    "google-auth",
    "requests",
    "urllib3",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langchain-google-vertexai>=2.1.2",
    "langchain-anthropic>=0.3.21",
    "langchain-community>=0.3.30",
    "langchain-xai>=0.2.5",
    "langchain-google-genai>=2.1.12",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.2",
]
